+++ src/mpack/mpack-expect.h
@@ -146,6 +146,7 @@ int32_t mpack_expect_i32(mpack_reader_t* reader);
  */
 int64_t mpack_expect_i64(mpack_reader_t* reader);
 
+#if MPACK_FLOAT
 /**
  * Reads a number, returning the value as a float. The underlying value can be an
  * integer, float or double; the value is converted to a float.
@@ -156,7 +157,9 @@ int64_t mpack_expect_i64(mpack_reader_t* reader);
  * @throws mpack_error_type if the underlying value is not a float, double or integer.
  */
 float mpack_expect_float(mpack_reader_t* reader);
+#endif
 
+#if MPACK_DOUBLE
 /**
  * Reads a number, returning the value as a double. The underlying value can be an
  * integer, float or double; the value is converted to a double.
@@ -167,7 +170,9 @@ float mpack_expect_float(mpack_reader_t* reader);
  * @throws mpack_error_type if the underlying value is not a float, double or integer.
  */
 double mpack_expect_double(mpack_reader_t* reader);
+#endif
 
+#if MPACK_FLOAT
 /**
  * Reads a float. The underlying value must be a float, not a double or an integer.
  * This ensures no loss of precision can occur.
@@ -175,7 +180,9 @@ double mpack_expect_double(mpack_reader_t* reader);
  * @throws mpack_error_type if the underlying value is not a float.
  */
 float mpack_expect_float_strict(mpack_reader_t* reader);
+#endif
 
+#if MPACK_DOUBLE
 /**
  * Reads a double. The underlying value must be a float or double, not an integer.
  * This ensures no loss of precision can occur.
@@ -183,6 +190,7 @@ float mpack_expect_float_strict(mpack_reader_t* reader);
  * @throws mpack_error_type if the underlying value is not a float or double.
  */
 double mpack_expect_double_strict(mpack_reader_t* reader);
+#endif
 
 /**
  * @}
@@ -429,6 +437,7 @@ MPACK_INLINE int mpack_expect_int_max(mpack_reader_t* reader, int max_value) {
     return mpack_expect_int_range(reader, 0, max_value);
 }
 
+#if MPACK_FLOAT
 /**
  * Reads a number, ensuring that it falls within the given range and returning
  * the value as a float. The underlying value can be an integer, float or
@@ -440,7 +449,9 @@ MPACK_INLINE int mpack_expect_int_max(mpack_reader_t* reader, int max_value) {
  * @throws mpack_error_type if the underlying value is not a float, double or integer.
  */
 float mpack_expect_float_range(mpack_reader_t* reader, float min_value, float max_value);
+#endif
 
+#if MPACK_DOUBLE
 /**
  * Reads a number, ensuring that it falls within the given range and returning
  * the value as a double. The underlying value can be an integer, float or
@@ -452,6 +463,7 @@ float mpack_expect_float_range(mpack_reader_t* reader, float min_value, float ma
  * @throws mpack_error_type if the underlying value is not a float, double or integer.
  */
 double mpack_expect_double_range(mpack_reader_t* reader, double min_value, double max_value);
+#endif
 
 /**
  * @}