+++ test/test-file.c
@@ -293,7 +293,7 @@ static void test_compare_print() {
     MPACK_FREE(actual_data);
 }
 
-#if MPACK_READER && MPACK_DEBUG
+#if MPACK_READER && MPACK_DEBUG && MPACK_DOUBLE
 static void test_print(void) {
 
     // miscellaneous print tests
@@ -326,7 +326,7 @@ static void test_print(void) {
 }
 #endif
 
-#if MPACK_NODE && MPACK_DEBUG
+#if MPACK_NODE && MPACK_DEBUG && MPACK_DOUBLE
 static void test_node_print(void) {
     mpack_tree_t tree;
 
@@ -955,10 +955,10 @@ void test_file(void) {
     FILE* blank = fopen(test_blank_filename, "wb");
     fclose(blank);
 
-    #if MPACK_READER && MPACK_DEBUG
+    #if MPACK_READER && MPACK_DEBUG && MPACK_DOUBLE
     test_print();
     #endif
-    #if MPACK_NODE && MPACK_DEBUG
+    #if MPACK_NODE && MPACK_DEBUG && MPACK_DOUBLE
     test_node_print();
     #endif